django__api/requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.1
defusedxml==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-bootstrap3==11.0.0
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-extensions==2.1.3
django-leaflet==0.24.0
django-sslserver==0.20
djangorestframework==3.8.2
djangorestframework-gis==0.13
GDAL==2.3.2
idna==2.7
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.6.1
pycparser==2.19
pyOpenSSL==18.0.0
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
ticketpy==1.1.2
Unidecode==1.0.22
urllib3==1.24.1
Werkzeug==0.14.1
